First pass to use expressjs-routes-info
