+++ lib/api/controllers/index.js
@@ -1,0 +1,11 @@
+const defaultViews = require('./default-views');
+const domain = require('./domain');
+const smnExamples = require('./smn-examples');
+const testData = require('./test-data');
+
+module.exports = {
+    defaultViews,
+    domain,
+    smnExamples,
+    testData
+};
