+++ package.json
@@ -30,6 +30,7 @@
     }
   },
   "dependencies": {
+    "@quoin/expressjs-routes-info": "0.1.3",
     "bluebird": "3.5.0",
     "lodash": "4.17.4",
     "mongodb": "2.2.25",
